+++ src/bomberman/entity/StaticEntity.java
@@ -9,21 +9,6 @@ package bomberman.entity;
  *
  * @author kdost
  */
-public class StaticEntity implements Entity {
-
-    @Override
-    public boolean isColliding(Entity b) {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-
-    @Override
-    public void draw() {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-
-    @Override
-    public void removeFromScene() {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
+public interface StaticEntity extends Entity {
     
 }
