+++ src/bomberman/entity/staticobjects/Wall.java
@@ -3,15 +3,17 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package bomberman.entity;
+package bomberman.entity.staticobjects;
 
+import bomberman.entity.Entity;
+import bomberman.entity.StaticEntity;
 import javafx.scene.paint.Color;
 
 /**
  *
  * @author kdost
  */
-public class Wall extends StaticEntity {
+public class Wall implements StaticEntity {
     private int height;
     private int width;
     private int health;
@@ -22,6 +24,21 @@ public class Wall extends StaticEntity {
     }
     public void getDamageLevel() {
         
+    }
+
+    @Override
+    public boolean isColliding(Entity b) {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
+    @Override
+    public void draw() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
+    @Override
+    public void removeFromScene() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
     
 }
